requirements.txt
fastapi
uvicorn
supabase
python-dotenv
openai
pypdf
streamlit
python-multipart
requests
